Move ntq_multi_stitch_blocks to stitch_blocks
